+++ src/figure_agent/renderers/__init__.py
@@ -1,0 +1,15 @@
+"""Renderer layer for ontology graphs."""
+
+from figure_agent.renderers.base import RenderConfig, RenderResult, Renderer
+from figure_agent.renderers.exceptions import LayoutError, RenderError, SVGRenderError
+from figure_agent.renderers.svg_renderer import SVGRenderer
+
+__all__ = [
+    "LayoutError",
+    "RenderConfig",
+    "RenderError",
+    "RenderResult",
+    "Renderer",
+    "SVGRenderError",
+    "SVGRenderer",
+]